rename indent -> format and add tests
