+++ packages/build-code/tests/format.test.ts
@@ -1,6 +1,10 @@
-import { getIndentation, indent } from "../src/indent";
+import { getIndentation, indent, removeLinesIncluding } from "../src/format";
 
 describe("getIndentation", () => {
+  it("works for no lines", () => {
+    expect(getIndentation("", "mySearch")).toEqual(0);
+  });
+
   it("works for one line", () => {
     const indentation = getIndentation("   mySearch", "mySearch");
     expect(indentation).toEqual(3);
@@ -8,9 +12,7 @@ describe("getIndentation", () => {
 
   it("works for multi-line", () => {
     const indentation = getIndentation(
-      `a bunch of stuff here
-
-    mySearch`,
+      "a bunch of stuff here\n\n    mySearch",
       "mySearch"
     );
 
@@ -25,12 +27,13 @@ describe("indent", () => {
   });
 
   it("works for multi-line", () => {
-    const indentation = indent(
-      `1
-2
-3abc  de `,
-      2
-    );
+    const indentation = indent("1\n2\n3abc  de ", 2);
     expect(indentation).toEqual("  1\n  2\n  3abc  de ");
   });
+});
+
+describe("removeLinesIncluding", () => {
+  expect(removeLinesIncluding("\n1\n   my text  \n2", "my text")).toEqual(
+    "\n1\n2"
+  );
 });
